Implement fee tx check in the trade protocol
After the deposit tx have been confirmed we check for the fee payment. For the buyer this is still not enough as we want to avoid that the buyer sends the payment and not only fail after he sent.
We will need to listen on the deposit tx confirmed state and then do the fee check and move to buyers step 2 only if the fee was paid and show otherwise a warning.