In MailboxMessageService we called requestPreliminaryData on requestDataManager at onTorNodeReady but that is called already on the P2PService. We remove the SetupListener from MailboxMessageService.
We do not return a boolean in requestPreliminaryData in case there is no seed
node (e.g. if we have setup only 1 seed) but call the onNoSeedNodeAvailable
at out listener instead.

At onTorNodeReady in P2PService we do not use the return value from
requestPreliminaryData to set isBootstrapped and call listeners for
onNoSeedNodeAvailable but we use the onNoSeedNodeAvailable which is
now called from the requestPreliminaryData method.

At mailboxMessageService we use the onBootstrapped method for both events
onUpdatedDataReceived and onNoSeedNodeAvailable.

In a future commit we will add a onBootstrapped method to our listener
as that is usually what clients are interested in so they can avoid the
boilerplate to listen on both events and execute the same code.

This issue that we called 2 times requestPreliminaryData at the same
seed was probably responsible for connectivity issues.