Add onBootStrapped method to BootstrapListener. Call onBootStrapped at BootstrapListener in P2PService from onUpdatedDataReceived or onNoSeedNodeAvailable.
There is some risk that the order of execution might be
different and could cause issues.
Further we call now also from onNoSeedNodeAvailable,
which might change behaviour in some clients.
It is a more correct implementation as handling for
onNoSeedNodeAvailable cases was omitted in most clients,
but it would require good testing and more thought if that change is safe.